# Tax Categories on Line Items respect updates to Variant and Product Tax Categories