--- a/core/app/models/spree/variant.rb
+++ b/core/app/models/spree/variant.rb
@@ -35,7 +35,7 @@ module Spree
     delegate :name, :description, :slug, :available_on, :discontinue_on, :discontinued?,
              :meta_description, :meta_keywords,
              to: :product
-    delegate :tax_category, to: :product, prefix: true
+    delegate :tax_category, :tax_category_id, to: :product, prefix: true
     delegate :shipping_category, :shipping_category_id,
       to: :product, prefix: true
     delegate :tax_rates, to: :tax_category
@@ -148,6 +148,15 @@ module Spree
     #
     def tax_category
       super || product_tax_category
+    end
+
+    # @return [Integer] the variant's tax category ID
+    #
+    # This returns the product's tax category ID if the tax category ID on the variant is nil. It looks
+    # like an association, but really is an override.
+    #
+    def tax_category_id
+      super || product_tax_category_id
     end
 
     # @return [Spree::ShippingCategory] the variant's shipping category