--- a/core/spec/models/spree/line_item_spec.rb
+++ b/core/spec/models/spree/line_item_spec.rb
@@ -5,6 +5,7 @@ require 'rails_helper'
 RSpec.describe Spree::LineItem, type: :model do
   let(:order) { create :order_with_line_items, line_items_count: 1 }
   let(:line_item) { order.line_items.first }
+  let(:target_shipment) { Spree::Shipment.new }
 
   context '#destroy' do
     it "fetches soft-deleted products" do
@@ -18,7 +19,8 @@ RSpec.describe Spree::LineItem, type: :model do
     end
 
     it "returns inventory when a line item is destroyed" do
-      expect_any_instance_of(Spree::OrderInventory).to receive(:verify)
+      line_item.target_shipment = target_shipment
+      expect_any_instance_of(Spree::OrderInventory).to receive(:verify).with(target_shipment)
       line_item.destroy
     end
 
@@ -30,8 +32,8 @@ RSpec.describe Spree::LineItem, type: :model do
   context "#save" do
     context "target_shipment is provided" do
       it "verifies inventory" do
-        line_item.target_shipment = Spree::Shipment.new
-        expect_any_instance_of(Spree::OrderInventory).to receive(:verify)
+        line_item.target_shipment = target_shipment
+        expect_any_instance_of(Spree::OrderInventory).to receive(:verify).with(target_shipment)
         line_item.save
       end
     end