--- a/core/spec/models/spree/tax/tax_helpers_spec.rb
+++ b/core/spec/models/spree/tax/tax_helpers_spec.rb
@@ -16,7 +16,8 @@ RSpec.describe Spree::Tax::TaxHelpers do
   let(:tax_category) { create(:tax_category) }
   let(:irrelevant_tax_category) { create(:tax_category) }
 
-  let(:item) { create(:line_item, tax_category:) }
+  let(:item) { create(:line_item, variant:) }
+  let(:variant) { create(:variant, tax_category:) }
   let(:tax_address) { item.order.tax_address }
   let(:zone) { create(:zone, name: "Country Zone", countries: [tax_address.country]) }
 
@@ -24,9 +25,11 @@ RSpec.describe Spree::Tax::TaxHelpers do
     create(:tax_rate, tax_categories: [tax_category], zone:)
   end
 
+  subject { DummyClass.new.valid_rates(item) }
+
   describe '#rates_for_item' do
     it 'returns tax rates that match the tax category of the given item' do
-      expect(DummyClass.new.valid_rates(item)).to contain_exactly(tax_rate)
+      expect(subject).to contain_exactly(tax_rate)
     end
 
     context 'when multiple rates exist that are currently not valid' do
@@ -41,7 +44,40 @@ RSpec.describe Spree::Tax::TaxHelpers do
       it 'returns only active rates that match the tax category of given item' do
         expect(Spree::TaxRate.for_address(tax_address)).to contain_exactly(tax_rate, invalid_tax_rate)
 
-        expect(DummyClass.new.valid_rates(item)).to contain_exactly(tax_rate)
+        expect(subject).to contain_exactly(tax_rate)
+      end
+    end
+
+    context "when the line_item's variant's tax_category is changed" do
+      let(:new_tax_category) { create(:tax_category) }
+      before do
+        variant.update(tax_category: new_tax_category)
+        tax_rate.update(tax_categories: [new_tax_category])
+      end
+
+      it "returns the new tax rate for the variant's tax category" do
+        expect(subject).to contain_exactly(tax_rate)
+      end
+    end
+
+    context "when item is a shipping_rate" do
+      let(:item) { create(:shipping_rate, shipping_method:) }
+      let(:shipping_method) { create(:shipping_method, tax_category:) }
+
+      it "returns the tax rate for the shipping_method's tax category" do
+        expect(subject).to contain_exactly(tax_rate)
+      end
+
+      context "when the shipping_method's tax_category is changed" do
+        let(:new_tax_category) { create(:tax_category) }
+        before do
+          shipping_method.update(tax_category: new_tax_category)
+          tax_rate.update(tax_categories: [new_tax_category])
+        end
+
+        it "returns the new tax rate for the shipping_method's tax category" do
+          expect(subject).to contain_exactly(tax_rate)
+        end
       end
     end
   end