@tvdeyen 2025-01-15
Thanks! Amazing PR. A great example for how open source contributions should be made. Super Good 👍🏻 

(unrelated linting failures will be fixed in a subsequent PR)